requirements.txt
openai
mutagen
python-dotenv
